Separate application endpoint to public and admin
