+++ admissions/admissions/tests/test_public_application.py
@@ -1,0 +1,105 @@
+from django.urls import reverse
+from rest_framework import status
+from rest_framework.test import APITestCase
+
+from admissions.admissions.models import Group, LegoUser, UserApplication
+from admissions.admissions.tests.utils import DEFAULT_ADMISSION_SLUG, create_admission
+
+
+class CreateApplicationTestCase(APITestCase):
+    def setUp(self):
+        global DEFAULT_ADMISSION_SLUG
+        self.admission_slug = DEFAULT_ADMISSION_SLUG
+        # Create admission and group
+        self.admission = create_admission()
+        self.webkom = Group.objects.create(name="Webkom")
+        self.koskom = Group.objects.create(name="Koskom")
+
+        # Setup Anna
+        self.pleb_anna = LegoUser.objects.create(username="Anna")
+
+        self.application_data = {
+            "text": "Ønsker Webkom mest",
+            "phone_number": "12345678",
+            "header_fields_response": {},
+            "applications": {
+                "webkom": "Hohohohohohohohohohooho webbis",
+                "koskom": "Hahahahahahahahahahaha arris",
+            },
+        }
+
+        # Setup Bob
+        self.pleb_bob = LegoUser.objects.create(username="Bob")
+
+    def test_cannot_apply_for_someone_else(self):
+        # Login as Bob, and try to apply as Anna. User should then be Bob
+        self.client.force_authenticate(user=self.pleb_bob)
+
+        annas_application_data = self.application_data.copy()
+
+        annas_application_data["user"] = self.pleb_anna.pk
+
+        res = self.client.post(
+            reverse(
+                "userapplication-list", kwargs={"admission_slug": self.admission_slug}
+            ),
+            annas_application_data,
+            format="json",
+        )
+
+        self.assertEqual(res.status_code, status.HTTP_201_CREATED)
+
+        pk_res = res.json()["pk"]
+
+        # Application registered as bob
+        self.assertEqual(UserApplication.objects.get(pk=pk_res).user, self.pleb_bob)
+
+    def test_can_apply(self):
+        self.client.force_authenticate(user=self.pleb_anna)
+        res = self.client.post(
+            reverse(
+                "userapplication-list", kwargs={"admission_slug": self.admission_slug}
+            ),
+            self.application_data,
+            format="json",
+        )
+
+        self.assertEqual(res.status_code, status.HTTP_201_CREATED)
+
+    def test_editing_application_works_correctly(self):
+        self.client.force_authenticate(user=self.pleb_anna)
+
+        # Apply first with webkom and koskom
+        self.client.post(
+            reverse(
+                "userapplication-list", kwargs={"admission_slug": self.admission_slug}
+            ),
+            self.application_data,
+            format="json",
+        )
+
+        self.assertEqual(
+            2,
+            UserApplication.objects.get(user=self.pleb_anna).group_applications.count(),
+        )
+
+        self.application_data = {
+            "text": "Ønsker Webkom mest",
+            "phone_number": "12345678",
+            "header_fields_response": {},
+            "applications": {"webkom": "Hohohohohohohohohohooho webbis"},
+        }
+
+        # Apply then only with webkom, removing koskom
+        self.client.post(
+            reverse(
+                "userapplication-list", kwargs={"admission_slug": self.admission_slug}
+            ),
+            self.application_data,
+            format="json",
+        )
+
+        self.assertEqual(
+            1,
+            UserApplication.objects.get(user=self.pleb_anna).group_applications.count(),
+        )
