+++ frontend/src/query/hooks.ts
@@ -2,22 +2,7 @@ import { useQuery } from "@tanstack/react-query";
 import { AxiosError } from "axios";
 import { Admission, Application, Group } from "src/types";
 
-// Admin hooks
-
-export const useAdminAdmissions = () => {
-  return useQuery<Admission[], AxiosError>({
-    queryKey: ["/admin/admission/"],
-  });
-};
-
-export const useAdminAdmission = (slug: string, enabled = true) => {
-  return useQuery<Admission, AxiosError>({
-    queryKey: [`/admin/admission/${slug}/`],
-    enabled,
-  });
-};
-
-// User hooks
+// Public hooks
 
 export const useAdmissions = () => {
   return useQuery<Admission[], AxiosError>({
@@ -31,12 +16,6 @@ export const useAdmission = (slug: string) => {
   });
 };
 
-export const useApplications = (slug: string) => {
-  return useQuery<Application[], AxiosError>({
-    queryKey: [`/admission/${slug}/application/`],
-  });
-};
-
 export const useMyApplication = (slug: string) => {
   return useQuery<Application, AxiosError>({
     queryKey: [`/admission/${slug}/application/mine/`],
@@ -46,5 +25,26 @@ export const useMyApplication = (slug: string) => {
 export const useGroups = () => {
   return useQuery<Group[], AxiosError>({
     queryKey: ["/group/"],
+  });
+};
+
+// Admin hooks
+
+export const useAdminApplications = (admissionSlug: string) => {
+  return useQuery<Application[], AxiosError>({
+    queryKey: [`/admission/${admissionSlug}/admin/application/`],
+  });
+};
+
+// Manage hooks
+
+export const useAdminAdmissions = () => {
+  return useQuery<Admission[], AxiosError>({ queryKey: ["/admin/admission/"] });
+};
+
+export const useAdminAdmission = (slug: string, enabled = true) => {
+  return useQuery<Admission, AxiosError>({
+    queryKey: [`/admin/admission/${slug}/`],
+    enabled,
   });
 };
